1. Merge branch 'make_homogen_table' into convert_homogentable 2. update ci 3. add a new parameters for HomogenTable class
