+++ dev/tools/check-gpu-cpu/run.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
-APP_NAME=check-gpu-cpu
+APP_NAME=list-compute-devices
 
 mpirun -n 1 ./${APP_NAME} false
