+++ mllib-dal/test.sh
@@ -112,7 +112,8 @@ source $OAP_MLLIB_ROOT/RELEASE
 
 export SPARK_VERSION=${SPARK_OPT:-$SPARK_VERSION}
 export PLATFORM_PROFILE=${PLATFORM_OPT:-$PLATFORM_PROFILE}
-export SPARK_CI_TEST=$TEST_OPT
+
+echo "computeDevice : $DEVICE_OPT"
 
 echo "computeDevice : $DEVICE_OPT"
